requirements.txt
certifi==2020.4.5.2
chardet==3.0.4
click==7.1.2
configparser==5.0.0
cycler==0.10.0
docker-pycreds==0.4.0
gitdb==4.0.5
GitPython==3.1.3
gql==0.2.0
graphql-core==1.1
idna==2.9
kiwisolver==1.2.0
matplotlib==3.2.2
mkl-fft==1.1.0
mkl-random==1.1.1
mkl-service==2.3.0
numpy==1.18.1
nvidia-ml-py3==7.352.0
olefile==0.46
pathtools==0.1.2
Pillow==7.1.2
promise==2.3
psutil==5.7.0
pyparsing==2.4.7
python-dateutil==2.8.1
PyYAML==5.3.1
requests==2.24.0
sentry-sdk==0.15.1
shortuuid==1.0.1
six==1.15.0
smmap==3.0.4
subprocess32==3.5.4
torch==1.5.1
torchvision==0.6.0a0+35d732a
tqdm==4.46.1
urllib3==1.25.9
wandb==0.9.1
watchdog==0.10.2
